# openapi.yml
openapi: 3.0.2
info:
  title: Structures service
  version: '0.1'
paths:
  /organization-units/{id}:
    get:
      tags:
        - Organization Units
        - Organization Units
      summary: Get Organization Unit
      operationId: get_organization_unit_organization_units__id__get
      parameters:
        - required: true
          schema:
            title: Id
            type: string
          name: id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationUnit'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
    put:
      tags:
        - Organization Units
        - Organization Units
      summary: Update Organization Unit
      operationId: update_organization_unit_organization_units__id__put
      parameters:
        - required: true
          schema:
            title: Id
            type: string
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUnitUpdateDto'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationUnit'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
    delete:
      tags:
        - Organization Units
        - Organization Units
      summary: Delete Organization Unit
      operationId: delete_organization_unit_organization_units__id__delete
      parameters:
        - required: true
          schema:
            title: Id
            type: string
          name: id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /organization-units:
    get:
      tags:
        - Organization Units
        - Organization Units
      summary: Find Organization Units
      operationId: find_organization_units_organization_units_get
      parameters:
        - required: false
          schema:
            title: Active
            type: boolean
          name: active
          in: query
        - required: false
          schema:
            title: Child Of
            type: string
          name: child_of
          in: query
        - required: false
          schema:
            title: Page
            type: integer
            default: 1
          name: page
          in: query
        - required: false
          schema:
            title: Limit
            type: integer
            default: 100
          name: limit
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationUnitPaginated'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
    post:
      tags:
        - Organization Units
        - Organization Units
      summary: Create Organization Unit
      operationId: create_organization_unit_organization_units_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUnitCreateDto'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationUnit'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /organization-units/{id}/activate:
    patch:
      tags:
        - Organization Units
        - Organization Units
      summary: Activate Organization Unit
      operationId: activate_organization_unit_organization_units__id__activate_patch
      parameters:
        - required: true
          schema:
            title: Id
            type: string
          name: id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationUnit'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /organization-units/{id}/deactivate:
    patch:
      tags:
        - Organization Units
        - Organization Units
      summary: Deactivate Organization Unit
      operationId: deactivate_organization_unit_organization_units__id__deactivate_patch
      parameters:
        - required: true
          schema:
            title: Id
            type: string
          name: id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationUnit'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /organization-units/{id}/change-parent/{new_parent}:
    patch:
      tags:
        - Organization Units
        - Organization Units
      summary: Change Parent Organization Unit
      operationId: change_parent_organization_unit_organization_units__id__change_parent__new_parent__patch
      parameters:
        - required: true
          schema:
            title: Id
            type: string
          name: id
          in: path
        - required: true
          schema:
            title: New Parent
            type: string
          name: new_parent
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationUnit'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
...
  /devices/{device_id}:
    get:
      tags:
        - Devices
        - Devices
      summary: Get Device
      operationId: get_device_devices__device_id__get
      parameters:
        - required: true
          schema:
            title: Device Id
            type: string
          name: device_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
    put:
      tags:
        - Devices
        - Devices
      summary: Update Device
      operationId: update_device_devices__device_id__put
      parameters:
        - required: true
          schema:
            title: Device Id
            type: string
          name: device_id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdateDto'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
    delete:
      tags:
        - Devices
        - Devices
      summary: Delete Device
      operationId: delete_device_devices__device_id__delete
      parameters:
        - required: true
          schema:
            title: Device Id
            type: string
          name: device_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /devices:
    get:
      tags:
        - Devices
        - Devices
      summary: Find Devices
      operationId: find_devices_devices_get
      parameters:
        - required: false
          schema:
            title: Active
            type: boolean
          name: active
          in: query
        - required: false
          schema:
            title: Child Of Organization Unit
            type: string
          name: child_of_organization_unit
          in: query
        - required: false
          schema:
            title: Child Of Outlet
            type: string
          name: child_of_outlet
          in: query
        - required: false
          schema:
            title: Page
            type: integer
            default: 1
          name: page
          in: query
        - required: false
          schema:
            title: Limit
            type: integer
            default: 100
          name: limit
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevicePaginatedDto'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
    post:
      tags:
        - Devices
        - Devices
      summary: Create Device
      operationId: create_device_devices_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreateDto'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /devices/{device_id}/activate:
    patch:
      tags:
        - Devices
        - Devices
      summary: Activate Device
      operationId: activate_device_devices__device_id__activate_patch
      parameters:
        - required: true
          schema:
            title: Device Id
            type: string
          name: device_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /devices/{device_id}/deactivate:
    patch:
      tags:
        - Devices
        - Devices
      summary: Deactivate Outlet
      operationId: deactivate_outlet_devices__device_id__deactivate_patch
      parameters:
        - required: true
          schema:
            title: Device Id
            type: string
          name: device_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /devices/{device_id}/change-outlet/{new_outlet_id}:
    patch:
      tags:
        - Devices
        - Devices
      summary: Change Outlet
      operationId: change_outlet_devices__device_id__change_outlet__new_outlet_id__patch
      parameters:
        - required: true
          schema:
            title: Device Id
            type: string
          name: device_id
          in: path
        - required: true
          schema:
            title: New Outlet Id
            type: string
          name: new_outlet_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
...
components:
  schemas:
    Device:
      title: Device
      required:
        - id
        - license
        - active
        - outlet_id
        - materialized_path
      type: object
      properties:
        id:
          title: Id
          type: string
        license:
          title: License
          type: string
        active:
          title: Active
          type: boolean
        outlet_id:
          title: Outlet Id
          type: string
        materialized_path:
          title: Materialized Path
          type: array
          items:
            type: string
    DeviceBase:
      title: DeviceBase
      required:
        - id
        - license
        - active
        - outlet_id
      type: object
      properties:
        id:
          title: Id
          type: string
        license:
          title: License
          type: string
        active:
          title: Active
          type: boolean
        outlet_id:
          title: Outlet Id
          type: string
    DeviceCreateDto:
      title: DeviceCreateDto
      required:
        - license
        - active
        - outlet_id
      type: object
      properties:
        license:
          title: License
          type: string
        active:
          title: Active
          type: boolean
        outlet_id:
          title: Outlet Id
          type: string
    DevicePaginatedDto:
      title: DevicePaginatedDto
      required:
        - pagination
        - data
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        data:
          title: Data
          type: array
          items:
            $ref: '#/components/schemas/DeviceBase'
    DeviceUpdateDto:
      title: DeviceUpdateDto
      type: object
      properties:
        license:
          title: License
          type: string
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    OrganizationUnit:
      title: OrganizationUnit
      required:
        - id
        - name
        - inn
        - kpp
        - active
        - parent_organization_unit
        - filler
        - materialized_path
      type: object
      properties:
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        inn:
          title: Inn
          type: integer
        kpp:
          title: Kpp
          type: integer
        active:
          title: Active
          type: boolean
        parent_organization_unit:
          title: Parent Organization Unit
          type: string
        filler:
          title: Filler
          type: string
        materialized_path:
          title: Materialized Path
          type: array
          items:
            type: string
    OrganizationUnitBase:
      title: OrganizationUnitBase
      required:
        - id
        - name
        - inn
        - kpp
        - active
        - parent_organization_unit
        - filler
      type: object
      properties:
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        inn:
          title: Inn
          type: integer
        kpp:
          title: Kpp
          type: integer
        active:
          title: Active
          type: boolean
        parent_organization_unit:
          title: Parent Organization Unit
          type: string
        filler:
          title: Filler
          type: string
    OrganizationUnitCreateDto:
      title: OrganizationUnitCreateDto
      required:
        - name
        - inn
        - kpp
        - legal_address
        - ogrn
        - filler
      type: object
      properties:
        parent_organization_unit:
          title: Parent Organization Unit
          type: string
        name:
          title: Name
          type: string
        inn:
          title: Inn
          type: integer
        kpp:
          title: Kpp
          type: integer
        legal_address:
          title: Legal Address
          type: string
        ogrn:
          title: Ogrn
          type: integer
        filler:
          title: Filler
          type: string
    OrganizationUnitPaginated:
      title: OrganizationUnitPaginated
      required:
        - pagination
        - data
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        data:
          title: Data
          type: array
          items:
            $ref: '#/components/schemas/OrganizationUnitBase'
    OrganizationUnitUpdateDto:
      title: OrganizationUnitUpdateDto
      required:
        - filler
      type: object
      properties:
        name:
          title: Name
          type: string
        inn:
          title: Inn
          type: integer
        kpp:
          title: Kpp
          type: integer
        legal_address:
          title: Legal Address
          type: string
        ogrn:
          title: Ogrn
          type: integer
        filler:
          title: Filler
          type: string
...
    Pagination:
      title: Pagination
      required:
        - page
        - limit
        - count
      type: object
      properties:
        page:
          title: Page
          type: integer
        limit:
          title: Limit
          type: integer
        count:
          title: Count
          type: integer
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
...
  securitySchemes:
    UserIdSchema:
      type: apiKey
      in: header
      name: X-User-Id